ADC: Correcting definitions of hardware registers from pointer to u32 to u8

+++ lib/ADC_private.h
@@ -1,10 +1,10 @@
 #ifndef _ADC_PRIVATE_H
 #define _ADC_PRIVATE_H
 
-#define ADMUX *((u32*) 0x27)
-#define ADCSRA *((u32*) 0x26)
-#define ADCH *((u32*) 0x25)
-#define ADCL *((u32*) 0x24)
+#define ADMUX *((u8*) 0x27)
+#define ADCSRA *((u8*) 0x26)
+#define ADCH *((u8*) 0x25)
+#define ADCL *((u8*) 0x24)
 
 
 #endif
